requirements.txt
flask
gunicorn
python-dotenv
flask-sqlalchemy
requests
